lint and prettify code

+++ contracts/interfaces/IAliumCallee.sol
@@ -2,5 +2,10 @@
 pragma solidity ^0.6.2;
 
 interface IAliumCallee {
-    function aliumCall(address sender, uint amount0, uint amount1, bytes calldata data) external;
+    function aliumCall(
+        address sender,
+        uint256 amount0,
+        uint256 amount1,
+        bytes calldata data
+    ) external;
 }
